requirements.txt
aiofiles==24.1.0
airtable==0.4.8
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Flask==3.1.2
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
priority==2.0.0
pyairtable==3.2.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
python-telegram-bot==22.4
Quart==0.20.0
requests==2.32.5
schedule==1.2.2
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
